Experimental: Reverse/Undo change log
In thinking deeply about how undo and rewind should work, it became clear that rewinding to any previous point in the past is going to depend on where undo is requested. For instance, if undo is requested at seq 999 out of 1000 then reconstructing the object from nothing requires re-merging in 999 diffs. One way to reduce that is to periodically take a snapshot, but this complicates the logic with additional storage. Instead, this diff introduces capturing an undo difference for every change. This changes the cost such that if we want to roll back just the last few messages, then the cost is proportional to the changes. This elevates the storage cost per transaction, but it also introduces some information redundancy to protect against cosmic radiation and bit flips.

testing: only validated no new breakage, this requires deeper testing