Update how-devkit-install.md
Grammar edits. Cleaned up some of the wording. Also, linked Java and Maven just to make it more convenient for users to quickly download and install the files.

As a side note, I see that the demo has a static username and password. Perhaps it might be necessary for security for them to change it if installed on production? Not sure if that is necessary for a demo, but it's something I thought of as I was reading it.