prototype: game schema reflection
This enables the devkit to reflect the schema to the browser via the websocket; the browser can now query the schema of the document such that it can generate client side code. This is essential for building a generic way of viewing tables, and it opens up the potential for adding metadata within types such that automatic editors/viewers can be forged.