contend with split brain at a low level
Instead of treating a sequencer off by one issue with a catastrophic failure, we leverage the new queue for documents with the undo log to (1) undo the recent execution, (2) download the head of the log, and then try again. There is a magic number built-in for the number of retries, and given this is a data race there is no back-off. Is this fine? well...