big decision - nuke time-driven-only state machines (for now)
This has to do with something that I could actually launch. As of now, I have no use-case for state machines that make progress on their own. Instead, I'll now require that someone is connected to the state machine for progress to be made based on time.

Afterall, when a tree falls in the woods, does it make a sound? Well, I can answer that question. No. It would be not be efficient to compute the tree falling sound file if no one is around