dynamic deployments are hard, going with static deployments
I had this idea of building this distributed deployer where each proxy host could decide to deploy to a specific host. This creates a thundering herd problem and a lot of redundant CPU. Instead, I'm binding a specific Adama host to a set of spaces via the.. database which it will scan on start up. Deployment now can pick a set of hosts to deploy, and we can build a future service to manage them