tune the db parameters as some connection limits were reached; need to investigate how RDS is setup
introduce a shell script to help tune the JVM during benchmarking. Also, add CLI support for listing documents. There is a problem when capacity shifts and the deployment table isn't updated. This creates dead space which only the overlord can rectify. We should thusly enforce timeouts on connect and the finder.