OK, fix a problem with the server Handler
If the load takes time (which it will), then we lose the disconnect signal! That's a huge problem, so  instead we craft a lazy core stream and associate a queue with it to buffer events until that load finishes. This fixed some odd behaviors while load testing, and made the graph look great!