kill the "undo" feature since it doesn't work
fix some coverage issues; it's like the tooling is now non-deterministic in coverage. So, I'm fixing the code to address this. I suspect this may have to do with some caching of bytecode which was invalidated with the removal of undo.

Also, I introduced the number of reads required to LocalDataChange to provide a signal to the DurableDocument who will need to schedule compaction tasks.