start thinking of how to bill for storage
storage is tricky because I want to sample it and then weight the monthly price, but I want to bill in pennies which requires tracking usage over time before a bill can be sent. That, or, I need to fractionate the pennies into smaller billable unit... a nano-penny? I could add another column for "byte-hours" which represents unbilled bytes. That may be OK to do.