almost near complete on the table journey
mod is now in the table and transforming to maybe<int> or maybe<long>; this is going to be obnoxious, BUT it makes division by zero impossible. This creates the opportunity to do some optimizations when a constant is on the right hand size.

Also, broke out the large math tables from BinaryExpression to lean it up.