get playful with the WAL
so, I can get 250 MB/second fairly reasonably.. I need to benchmark how much of that is the encoder vs the IO. However, this is a decent place to start. It's fairly low latency, and a key finding is that I need to rethink compact() and maybe axe it.

I also introduced arrays to the code generator tool for Messages, so that's pretty neat.