Introduce the assetId encoder
Assets ids are going to be considered private to Adama, and viewers per connection will need to provide an a key (via a header) to get access to the image. The mechanism of setting and getting the header are still in conceptualization, but it may be as simple as having a new API available.