start direct testing the crazy Machine class
This is going to require thinking about how to inject different failure modes, but this should be fun. Need to switch gears and get the caravan supporting archive. The key way to get archive working is to inline the sequencers and then just dump the stored list, this way we can then do a merge for duplicate backups