Introduce __hash on messages
This may be essential for deduping requests that are retries due to network failure. I can use this for RPC, @web put, and maybe @web get (if I think about it...). Although, the more I think about it, the weirder it may be. This can also be useful for general messages, but then messages can't be repeated it. I should really add some signals to make this automatic and not worry to much about it.