Figure out generational delta text sequencing
Since clients rely on a differential stream, the DText needs to know when to send an entire copy. The gen parameter allowed this to be inferred on set, but it failed when the entire backing RxText changed which had its own gen space. Here, we make all gen unique per RxText and set pairing by bringing an auto-incrementing field into play. This should allow a text object to be selected from a set of items.