Retry a race condition and minimize the impact of race condition
Retry on seq off which should be rare, but this caused me to find a rare race condition in the cache where multiple operations against the data store _could_ cause a race condition such that two items in the cache are loaded. We now measure this as waste, and use the appropriate operation. This handles a bug which may be more appropriate to handle within CoreService where creation/load are handled by a simpler queue rather than inflight operations as they may conflict. See #129