dynamic type extensions work and are much more useful in practice
With the monad chaining of maybe, I can now do something like if(props["x"].i() as x) { ... } which is expressive and results in a value only for the good things (x exists and is an integer type).

Also, hooked up extensions to a few other types so I'm not surprised later like I was today.