Start work on a new billing system
A complication of building the billing engine is the flexibility of adjusting to stripe. As such, I want to build it in Adama as Adama is pretty cool. This means I need to send billing records to Adama rather than clean up the database. Furthermore, I need to think about overlord in a multi-region context as I have a new data-center spinning up as a single-node.