start the process of introducing timeouts to fetches
This lays groundwork for storing timeouts so some operations can get on with their life. This introduces a rough sketch of the timeout data structures with appropriate unit tests. The data structures are put into place in key areas, but final integration will require a more polished inspection of LivingDocument. The crazy bit is regenerating the tests.