Introduce a KeyMap to DurableListStore
So, I'm preparing for a one-way door where DLS is responsible for mapping keys to ids; this will require production to be archived prior to deployment. This diff is benign as it doesn't introduce anything beyond reading the keymap which isn't persisted yet.

For safety reasons, I plan to enable a smooth migration such that I can start building the keymap and then I need a way to validate that production has drained before a cutover.