prototype of localmode
A big chunk that is missing is third party services. This is because third party services requires... keys. Keys are stored encrypted, so we will need a good way to translate those keys locally WITHOUT decryption. This could be done via an override of sorts which I'll start but... there is the trust thing.