So, I was wrong...
Long story short, I don't want another API... It's going to be a lot of work for little bang. Instead, I need to fix the old path with multiple facets and then make the global/region boundary more clear within saas. I'm introducing scopes that allow me to indicate at the API level which are global/region. This doesn't address some of the boundaries, but it helps with a lot of the API issues.