self.IT doesn't need anything more fancy that IF
At core, we have to fork on an iterator pass simply because you can escape to a higher level. This practically requires to think about some optimization work so and delay subscribes while an update is taking place such that we can batch the updates. However, that is a TODO for later times.