findbind was a bad idea
Ok, so, using findbind was a diaster that took a bit recover from. Basically, bind() is very special in the state machine process. Without the bind() taking place, it thought it was already on the machine and so no restore was needed. This meant documents were in a half-baked state and unable to make progress because caravan didn't know about them.