two items
First, post and domain based connections were not routing for adama.games; this is tricky because I don't want to invent a config for how it should work. Instead, adama.games uses "default-document" as the key for both post and creating domain connections.

Also, updating the viewer state no longer calls invalidation, so this will reduce a bunch of chatter.

tests are running right now