the whole split between devbox services and production services is problematic
this starts the path of having a unified path where the exceptions are the norm, and this is being done by introducing a ServiceConfigFactory and making the ServiceConfig an interface where there will be a devbox version and a production version. This also has the potential of enabling unit testing the remote service as well.