Fix a reactivity bug and clean up invalidation logic
Context: A global bubble (B1) was depending on a table (T) while a record in that table had a bubble (B2) that didn't depend on the object being mutated and instead relied on a global map (M). When M updated, B2 got invalidated but the B1 didn't get updated as T and B2 were disconnected. This patch establishes that a bubble will invalidate the hosting parent (so, B2 will invalidate T which will invalidate B1). As part of this, invalidation deduping is becoming a consistent idiom within the various RxParent implementations which may also improve performance during high reactive storms.