start plumbing traffic hints up the stack
The traffic hint is going to be an encoded instruction indicating how to handle traffic. At this time, the signal comes from the document and this will be transformed in a variety of ways. The important thing is that this is effectively a special opaque header defining an open contract between the LB and Adama to leverage.